Add a base class
