+++ services/spiget/spiget-downloads.service.js
@@ -1,14 +1,9 @@
 'use strict'
 
-const BaseJsonService = require('../base-json')
+const { BaseSpigetService, documentation } = require('./spiget-base')
 const { metric } = require('../../lib/text-formatters')
 
-const Joi = require('joi')
-const schema = Joi.object({
-  downloads: Joi.number().required(),
-}).required()
-
-module.exports = class SpigetDownloads extends BaseJsonService {
+module.exports = class SpigetDownloads extends BaseSpigetService {
   static get route() {
     return {
       base: 'spiget/downloads',
@@ -28,24 +23,12 @@ module.exports = class SpigetDownloads extends BaseJsonService {
     return this.constructor.render({ downloads })
   }
 
-  async fetch({ resourceid }) {
-    const url = `https://api.spiget.org/v2/resources/${resourceid}`
-
-    return this._requestJson({
-      schema,
-      url,
-    })
-  }
-
   static render({ downloads }) {
     return {
       message: metric(downloads),
     }
   }
 
-  static get category() {
-    return 'other'
-  }
   static get examples() {
     return [
       {
@@ -54,6 +37,7 @@ module.exports = class SpigetDownloads extends BaseJsonService {
           resourceid: '9089',
         },
         staticPreview: this.render({ downloads: 560891 }),
+        documentation,
       },
     ]
   }
